Switch to using python-dotenv
Ideological discussions about 12-factor apps aside, we currently have
two deployment-specific secrets (DB URI and CSRF token) and we're
assuredly going to add more later. We need a way of managing those that
isn't "creating and managing a new file with secrets" or "stuffing them
in the global environment".

Enter dotenv. This python module lets you specify things in a file
called .env that then get added to the os.environ namespace. This is
handy because instead of deployment requiring hand-hacking of the
process environment, it just lives with the app. It's also a good
starting point if future users of OO don't want to janitor their own
servers: heroku, for instance, mostly works through .env files.

This change did some reworking to the app itself that we should test out
before sending to prod; it also changes the provisioning of the vagrant
VM to properly plop the right thing into .env. Now the app starts and
runs with sample data automatically in a vagrant. I also made the
create_db.py script idempotent-ish, so it doesn't do anything if it
receives a DatabaseAlreadyControlledError.

I've also updated the dev and deploy documentation to reflect our wonderful
New Reality.